Add support for multiple installations
This adds the flags `install-id` and `owner` that can be used to identify
an installation to be used for authentication.

The `install-id` flag accepts an installation id that can be found with
our REST API [List installations for authenticated
app](https://docs.github.com/en/rest/reference/apps#list-installations-for-the-authenticated-app)

The `owner` flag accepts an organization or user to identify an
installation to be used for authentication.

By default the last installation is used.